% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firing.R
\name{firing.rate}
\alias{firing.rate}
\title{Calculate the Firing Rate of a Neuron or Neurons given their Spiketrains}
\usage{
firing.rate(neuronID, spiketrain)
}
\arguments{
\item{neuronID}{number}

\item{spiketrain}{data frame}
}
\value{
a number indicating the firing rate in spikes per frame
}
\description{
Calculate the Firing Rate of a Neuron or Neurons given their Spiketrains
}
